remove unused controller
